requirements.txt
torch 
torchvision 
torchaudio
opencv-python==4.5.3.56
Matplotlib
openmim
mmcv-full
mmdet